create and update working fine in reservations

+++ database/migrations/2024_05_26_180826_create_tables_table.php
@@ -11,7 +11,7 @@ return new class extends Migration
         Schema::create('tables', function (Blueprint $table) {
             $table->id('table_id');
             $table->string('number');
-            $table->unsignedInteger('capacity');
+            $table->unsignedTinyInteger('capacity');
             $table->timestamps();
         });
     }
